laravel6

+++ database/migrations/2020_09_30_182339_create_produtos_table.php
@@ -17,7 +17,7 @@ class CreateProdutosTable extends Migration
             $table->id();
             $table->string('nome');
             $table->string('categoria');
-            $table->float('valor');
+            $table->double('valor');
             $table->string('imagem');
             $table->timestamps();
         });
